move : domain.common 내 클래스 정리
