+++ src/main/java/com/playus/twpservice/domain/chat/repository/ChatMessageRepository.java
@@ -1,7 +1,7 @@
 package com.playus.twpservice.domain.chat.repository;
 
 import com.playus.twpservice.domain.chat.entity.ChatMessage;
-import com.playus.twpservice.domain.common.BaseMongoRepository;
+import com.playus.twpservice.domain.common.data.BaseMongoRepository;
 import org.springframework.data.mongodb.repository.Query;
 import org.springframework.data.mongodb.repository.Update;
