+++ src/main/java/com/playus/twpservice/domain/chat/repository/ChatRoomRepository.java
@@ -1,7 +1,7 @@
 package com.playus.twpservice.domain.chat.repository;
 
 import com.playus.twpservice.domain.chat.entity.ChatRoom;
-import com.playus.twpservice.domain.common.BaseMongoRepository;
+import com.playus.twpservice.domain.common.data.BaseMongoRepository;
 
 public interface ChatRoomRepository extends BaseMongoRepository<ChatRoom, String> {
