+++ src/main/java/com/playus/twpservice/domain/common/validation/Describable.java
@@ -1,4 +1,4 @@
-package com.playus.twpservice.domain.common;
+package com.playus.twpservice.domain.common.validation;
 
 public interface Describable {
     String getDescription();
