+++ src/main/java/com/playus/twpservice/domain/party/enums/PartyAgeGroup.java
@@ -1,6 +1,6 @@
 package com.playus.twpservice.domain.party.enums;
 
-import com.playus.twpservice.domain.common.Describable;
+import com.playus.twpservice.domain.common.validation.Describable;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
