+++ src/main/java/com/playus/twpservice/domain/party/repository/read/PartyReadOnlyRepository.java
@@ -1,9 +1,8 @@
 package com.playus.twpservice.domain.party.repository.read;
 
-import com.playus.twpservice.domain.common.BaseMongoRepository;
+import com.playus.twpservice.domain.common.data.BaseMongoRepository;
 import com.playus.twpservice.domain.party.document.PartyDocument;
 import com.playus.twpservice.domain.party.repository.read.custom.PartyReadOnlyRepositoryCustom;
-import org.springframework.data.mongodb.repository.MongoRepository;
 
 public interface PartyReadOnlyRepository extends BaseMongoRepository<PartyDocument, Long>, PartyReadOnlyRepositoryCustom {
