+++ src/main/java/com/playus/twpservice/domain/party/repository/read/PartyThumbnailUrlReadOnlyRepository.java
@@ -1,6 +1,6 @@
 package com.playus.twpservice.domain.party.repository.read;
 
-import com.playus.twpservice.domain.common.BaseMongoRepository;
+import com.playus.twpservice.domain.common.data.BaseMongoRepository;
 import com.playus.twpservice.domain.party.document.PartyThumbnailUrlDocument;
 
 public interface PartyThumbnailUrlReadOnlyRepository extends BaseMongoRepository<PartyThumbnailUrlDocument, Long> {
